+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/configuration/action/single/ReversibleAction.kt
@@ -1,8 +1,6 @@
 package com.badoo.ribs.core.routing.configuration.action.single
 
 import android.os.Parcelable
-import com.badoo.ribs.core.routing.configuration.ConfigurationContext
-import com.badoo.ribs.core.routing.transition.TransitionElement
 
 internal interface ReversibleAction<C : Parcelable> : Action<C> {
     fun reverse()
